backend/requirements.txt
fastapi
uvicorn
python-multipart
elasticsearch
pinecone
langchain-community
langchain-text-splitters
langchain-huggingface

sentence-transformers
pypdf
google-generativeai